syscall: Skip cleanup and ignore sysautoloadbottom
Since sysautoload is not configured, sysautoload is set
to /dev/null and sysautoloadbottom is set to /dev/null.bottom,
which is not a valid file.
This patch simply ignores any operations on sysautoloadbottom
to avoid unwanted behaviors.
An alternative workaround would be defining sysautoload in
kern/lib/libsys_linux/syscalls.conf, but I will not
investigate this issues further.

Signed-off-by: Seonghyun Park <shpark1@protonmail.com>